+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2024-01-18
+* Supports fetchEager API property
+
 ## 2023-12-11
 ### Fixed
 * Fixed pagination for queries without joins